Refactor optimizers into sub packages
