+++ src/main/java/com/facebook/openwifirrm/modules/ApiServer.java
@@ -44,13 +44,13 @@ import com.facebook.openwifirrm.RRMConfig.ModuleConfig.ApiServerParams;
 import com.facebook.openwifirrm.RRMConfig.ServiceConfig;
 import com.facebook.openwifirrm.Utils.LruCache;
 import com.facebook.openwifirrm.VersionProvider;
-import com.facebook.openwifirrm.optimizers.LeastUsedChannelOptimizer;
-import com.facebook.openwifirrm.optimizers.LocationBasedOptimalTPC;
-import com.facebook.openwifirrm.optimizers.MeasurementBasedApApTPC;
-import com.facebook.openwifirrm.optimizers.MeasurementBasedApClientTPC;
-import com.facebook.openwifirrm.optimizers.RandomChannelInitializer;
-import com.facebook.openwifirrm.optimizers.RandomTxPowerInitializer;
-import com.facebook.openwifirrm.optimizers.UnmanagedApAwareChannelOptimizer;
+import com.facebook.openwifirrm.optimizers.channel.LeastUsedChannelOptimizer;
+import com.facebook.openwifirrm.optimizers.channel.RandomChannelInitializer;
+import com.facebook.openwifirrm.optimizers.channel.UnmanagedApAwareChannelOptimizer;
+import com.facebook.openwifirrm.optimizers.tpc.LocationBasedOptimalTPC;
+import com.facebook.openwifirrm.optimizers.tpc.MeasurementBasedApApTPC;
+import com.facebook.openwifirrm.optimizers.tpc.MeasurementBasedApClientTPC;
+import com.facebook.openwifirrm.optimizers.tpc.RandomTxPowerInitializer;
 import com.facebook.openwifirrm.ucentral.UCentralClient;
 import com.facebook.openwifirrm.ucentral.UCentralUtils;
 import com.facebook.openwifirrm.ucentral.gw.models.SystemInfoResults;
